requirements.txt
python-dotenv~=1.0.0
config==0.5.1
requests~=2.31.0
